Added Avatar States And Improved Folder Structure
- Added Avatar Loading States
- Improved Helpers File/Folder Structure
