+++ src/lib/components/ui/accordion/AccordionItem.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import { cn } from '$lib/utils';
+	import { cn } from '$lib/helpers/utils';
 	import { getContext, setContext } from 'svelte';
 	import { writable, type Writable } from 'svelte/store';
