+++ src/lib/components/ui/alert/Alert.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import { cn } from '$lib/utils';
+	import { cn } from '$lib/helpers/utils';
 	import type { VariantProps } from 'class-variance-authority';
 	import { alertVariants } from './alertVariants';
