+++ src/lib/components/ui/avatar/Avatar.svelte
@@ -1,10 +1,11 @@
 <script lang="ts">
 	import { setContext } from 'svelte';
 	import { writable, type Writable } from 'svelte/store';
+	import type { AvatarLoadingtatus } from '.';
 
-	let imageLoadedStore: Writable<boolean> = writable();
+	let imageLoadingStatusStore: Writable<AvatarLoadingtatus> = writable('idle');
 
-	setContext('avatar', imageLoadedStore);
+	setContext('avatar', imageLoadingStatusStore);
 </script>
 
 <span class="relative flex h-10 w-10 shrink-0 overflow-hidden rounded-full">
