+++ src/lib/components/ui/avatar/AvatarImage.svelte
@@ -1,24 +1,26 @@
 <script lang="ts">
 	import { createEventDispatcher, getContext, onMount } from 'svelte';
 	import type { Writable } from 'svelte/store';
+	import type { AvatarLoadingtatus } from '.';
 
-	const dispatch = createEventDispatcher<{ loadingStatusChange: boolean }>();
+	const dispatch = createEventDispatcher<{ loadingStatusChange: AvatarLoadingtatus }>();
 
 	export let src: string;
 	export let alt: undefined | string = undefined;
 
-	let imageLoadedStore: Writable<boolean> = getContext('avatar');
+	let imageLoadingStatusStore: Writable<AvatarLoadingtatus> = getContext('avatar');
 
-	$: dispatch('loadingStatusChange', $imageLoadedStore);
+	$: dispatch('loadingStatusChange', $imageLoadingStatusStore);
 
 	onMount(() => {
+		imageLoadingStatusStore.set('loading');
 		const img = new Image();
 		img.src = src;
-		img.onload = () => imageLoadedStore.set(true);
-		img.onerror = () => imageLoadedStore.set(false);
+		img.onload = () => imageLoadingStatusStore.set('loaded');
+		img.onerror = () => imageLoadingStatusStore.set('error');
 	});
 </script>
 
-{#if $imageLoadedStore}
+{#if $imageLoadingStatusStore === 'loaded'}
 	<img {src} {alt} class="aspect-square h-full w-full" />
 {/if}
