+++ src/lib/components/ui/avatar/index.ts
@@ -1,3 +1,4 @@
-export { default as Avatar } from "./Avatar.svelte";
-export { default as AvatarFallback } from "./AvatarFallback.svelte";
-export { default as AvatarImage } from "./AvatarImage.svelte";
+export { default as Avatar } from './Avatar.svelte';
+export { default as AvatarFallback } from './AvatarFallback.svelte';
+export { default as AvatarImage } from './AvatarImage.svelte';
+export type AvatarLoadingtatus = 'idle' | 'loading'| 'loaded' | 'error';
