+++ src/lib/components/ui/input/Input.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import { cn } from '$lib/utils';
+	import { cn } from '$lib/helpers/utils';
 	import type { HTMLInputTypeAttribute } from 'svelte/elements';
 
 	let className: string | undefined | null = undefined;
