+++ src/lib/components/ui/separator/Separator.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import { cn } from '$lib/utils';
+	import { cn } from '$lib/helpers/utils';
 	import type { SeparatorOrientation } from '.';
 
 	let className: string | undefined | null = undefined;
