+++ src/lib/components/ui/switch/Switch.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import { cn } from '$lib/utils';
+	import { cn } from '$lib/helpers/utils';
 	import { createEventDispatcher } from 'svelte';
 
 	const dispatch = createEventDispatcher<{ checkedChange: boolean }>();
