+++ src/lib/components/ui/textarea/Textarea.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import { cn } from '$lib/utils';
+	import { cn } from '$lib/helpers/utils';
 
 	let className: string | undefined | null = undefined;
 	export { className as class };
