+++ src/lib/helpers/utils.ts
@@ -1,0 +1,19 @@
+import type { ClassValue } from 'clsx';
+import { clsx } from 'clsx';
+import { twMerge } from 'tailwind-merge';
+
+export function cn(...inputs: ClassValue[]): string {
+	return twMerge(clsx(inputs));
+}
+
+export function removeValue<T>(source: T[], target: T): T[] {
+	const index = source.indexOf(target);
+	if (index !== -1) {
+		source.splice(index, 1);
+	}
+	return source;
+}
+
+export function hasValue<T>(source: undefined | T | T[], target: T): boolean {
+	return (!Array.isArray(source) && source === target) || (Array.isArray(source) && source?.includes(target));
+}
